[Mojave Release Wk2] Layout Test http/tests/cache/disk-cache/disk-cache-validation-attachment.html is flaky https://bugs.webkit.org/show_bug.cgi?id=190323 <rdar://problem/45056253>
Reviewed by Youenn Fablet.

* http/tests/cache/disk-cache/disk-cache-validation-attachment-expected.txt:
* http/tests/cache/disk-cache/disk-cache-validation-attachment.html:
Simplify test a bit to only test the header combinations it cares about.

* http/tests/cache/disk-cache/resources/cache-test.js:
(runTestsNow):
(runTests):
The flakiness was due to logic in CachedResourceLoader::determineRevalidationPolicy() which more aggressively
reuses cached resources (bypassing freshness checks) until the load event has fired. This means that the
test would give different results depending on whether or not the load event had a chance to fire before
we do the checks. To address the issue, update runTests() to wait for the load event if it has not yet
fired. This will likely address flakiness in other disk-cache tests as well.

* platform/mac-wk2/TestExpectations:
Unskip test that is no longer flaky.


Identifier: 207216@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239137 268f45cc-cd09-0410-ab3c-d52691b4dbfc