Unparented WKWebView can't retrieve main resource data for a main frame plugin https://bugs.webkit.org/show_bug.cgi?id=192923 <rdar://problem/46859068>
Reviewed by Wenson Hsieh.

Source/WebKit:

* WebProcess/Plugins/PluginView.cpp:
(WebKit::PluginView::liveResourceData const):
If the WKWebView is unparented, we may not immediately initialize the plugin.
In that case, PluginView holds on to the backing data until the plugin
is initialized. If a WKWebView API client asks for the backing data
for the plugin during this time, we should return it, instead of bailing.

Tools:

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/WebKitCocoa/GetResourceData.mm:
Add a test ensuring that both parented and unparented WKWebViews can
retrieve main resource data when loading a PDF.


Identifier: 207483@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239442 268f45cc-cd09-0410-ab3c-d52691b4dbfc